requirements.txt
httpx>=0.24.0
asyncio
PyYAML>=6.0 